Corectat functia move in character, adaugat functiile getIntPosition si getFloatPosition

+++ source/Game.cpp
@@ -41,7 +41,7 @@ bool Game::Init(const std::string& title, int x, int y, int width, int height, i
     m_map = new Map(30, 30);
     m_map->createRandomMap();
 
-    Character* character = new Character(0, 0, "Nume", std::make_pair<int, int>(0, 0), new Weapon());
+    Character* character = new Character(0, 0, "Nume", std::make_pair<float, float>(0, 0), new Weapon());
     m_characters.push_back(character);
 
     m_running = true;
